pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "maildigger"
version = "0.1.0"
description = "Dig Gmail emails and attachments into LLM-friendly markdown"
requires-python = ">=3.10"
dependencies = [
    "click>=8.1.0",
    "pdfplumber>=0.10.0",
    "python-docx>=1.0.0",
    "openpyxl>=3.1.0",
    "python-pptx>=0.6.23",
    "beautifulsoup4>=4.12.0",
    "markdownify>=0.11.0",
    "rich>=13.0.0",
]

[project.scripts]
maildigger = "maildigger.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
